test: reuse same abstract class in gradle tests
